Properly handle vector parameters in resolution.
Updated the neddoc plugin, the parameters view and the module hierarchy view.
Both single and vector parameters should be handled mostly correctly except
some relly weird cases. The concept of total parameter assignment was introduced
for vector parameters, so that better estimation of parameter assignments can be
achieved. In general this problem cannot be solved, consider a submodule vector
of size 2 with a parameter p. The parameter is unassigned at its declaration,
it is assigned at the compound module, but only index 0 computed by an expression.
In the ini file the index 1 is simply assigned to a value. There's no way to know
if there is a submodule for that parameter that is unassigned without actually
evaluating all expression.