IDE: workaround for NPE in inifile editor.
NPE occurred when opening a very large inifile for a very large project,
and it was timing dependent. It also occured with larger probability
when file had a syntax error. NPE occurred while refreshing the Outline
view.

TODO: it should be checked how we can get a null ptr here.

X-Committed-on-Branch: master