Scave: allow standalone "scave" program to be launched with command-line args
The new "autoimport" plugin interprets command-line args, and creates
projects and anf files automatically on startup to cover the files
specified on the command line.

Synopsis:
  scave VECFILES_SCAFILES_OR_DIRECTORIES
  scave ANF_FILE VECFILES_SCAFILES_OR_DIRECTORIES
  scave ANF_FILES

The first two cases create an ANF file for the result files and directories.
One may specify wildcards for files, e.g. *.vec or results/A*.sca.
For directories, DIR/*.vec and DIR/*.sca will be added.
The third case opens the given ANF files in the editor.

Also updated scave.product and related files.

X-Committed-on-Branch: master