Tkenv: fix for bug #195: broadcast animation crashes if msg gets deleted by the end of event processing
Note: current solution still has issues if a message is deleted
and a new message object is created at the same memory address,
as their animations will get mixed up. A workaround would be to
augment the message pointer with the message ID in anim_remember_msg
to disambiguate it, but the result would still not be perfect
(see parallel tictoc).

X-Committed-on-Branch: master