Fix for bug#227.
Added NPE guard to SetConstraintCommand (old bound is no longer a required attribute)
Adjusted the PIN location (it was off with half the height of the label, since the ISubmoduleConstaint was introduced)
Added UnsupportedOperation to getConstraint, to avoid future accidental calls.
The figure itself is the constraint object.

X-Committed-on-Branch: master