Inifile Editor: bug #132: analyze inifile in the background
Background thread analyzes a readonly copy of the inifile (eliminates the need for locking.)
Clients can specify a timeout how long they are willing to wait for the results, and/or get notified via an IAnalysisListener when it completes.
When something changes in the inifile or NedResources, analysis is re-scheduled after some small delay.

Also:
- option to disable/enable analysis (see context menu)
- build ModuleHierarchy tree lazily (to reduce refresh time for large trees)
- ModuleParametersView: update when analysis has completed

[Commit msg by Andras]

X-Committed-on-Branch: topic/inifile-features