Inifile editor: expressions are accepted in iterations, incl variable references (Tamas Borbely)
Validation currently just checks that referenced variables exist.
Expressions with $var references in them are not validated, because
we don't know what the contents of $var would be (we currently don't
follow variable references during validation).

Implementation of the same feature in the simulation runtime will come
in a separate commit.

X-Committed-on-Branch: master