opp_run: Changed how opp_run is used to start debug and release models.
There are two programs to start models built as shared libs:
- 'opp_run_release' is built using release opp libraries and starts any release
  model directly by invoking evMain.
- 'opp_run' is linked against debug opp libs. It first loads all libraries
  into the address space dynamically and check for the simbol __is_release_opp_sim
  (that is present only in release builds). If it does not find it then the
  model is debug and proceeds normally by invoking evMain. If it decides that
  the model is 'release' it reloads the whole simulation by exec-ing it.

The simulation IDE also detects whether release or debug version of the model is
started and it starts it with either opp_run or opp_run_release. This is required because
on windows opp_run_release is started as a different process and the IDE could not
control/kill the opp_run_release process if it would be started by opp_run.

X-Committed-on-Branch: topic/opp_run2