sim: cMessageHeap fix plus test case
Bug: on scheduling a message for the current simulation time: if there
was already a message scheduled EARLIER for the same simulation time,
the new message was still stored in the circbuf, therefore overtaking
the message scheduled earlier. This should not happen: message with the
same simulation time and priority should be delivered in the order they
were scheduled.

Fix: do not use circbuf if there's a message in the heap part scheduled
earlier for the current simulation time.

X-Committed-on-Branch: master