IDE: C++ toolchain fix (#407): The ScannerInfoConsoleParser removed. (and removed an earlier workaround for this issue)
We do not need to scan the build process output and parse DEFINES and INCLUDES from there.
All includes and defines are added correctly by hand based on the cproject description.
The parsing of output caused an interference with the macros defined by project features.
(After disabling a feature, the scanner added back the feature's macro on the next build process,
this led to inconsistencies in the configuration = disabled feature but the WITH_XXX macro present)

X-Committed-on-Branch: topic/depcache

X-Committed-on-Branch: master