DependencyCache: fundamental rewrite
Basic problem: when a Project Feature is turned on/off, the cached
per-file #include lists need to be invalidated due to affected
#ifdef WITH_xxx blocks. The old code did not do that.

Another fault of old implementation: using ITranslationUnit.getIncludes()
to get the list of active includes is SLOW, because CDT generally does
NOT keep translation units  in memory, so getIncludes() needs to parse
the file first. After a Feature change, all files have to be re-parsed,
which takes about as much time as indexing.

Additional, somewhat unrelated issue: when a Project Feature is
enabled/disabled, CDT does NOT automatially refresh the index (although
#ifdef enablements change).

Solution: use the index to get the list of active includes. Before that,
we need to ensure that the index is up to date (start the index update
process, and wait until it completes). One problem with the index is that
it does not store the line numbers of #includes; so when we want to add
a marker (error/warning) for an #include, we need to read the file from
disk and seach for the #include in it to figure out its line number.

The rewrite has significantly reduced the complexity of DependencyCache
as well. We can get the active #includes from the index quite fast, and
resolving them and computing per-file and per-folder dependencies are
also quite fast -- it all takes about 300-500ms on an INET-scale project.
Thus, we invalidate and recompute everything for all projects at once
(one cannot invalidate one project only). This change has greatly reduced
the complexity of DependencyCache.

The delayedJob that updates markers have been removed. It did not add
much to the IDE, and continually keeping the markers up to date
would mean that the index has to be up to date all the time -- which is
inconvenient when the IDE is used with a slow CPU.

New code: whenever a project C/C++ configuration (ICCProjectDescription)
changes, we assume that a Project Feature was enabled/disabled or some
other (significant) project reconfiguration took place. In response,
we invalidate the dependency cache and start a full reindexing of the
project (and the projects that depend on the affect project).

MakefileBuilder change: adapted to change that getProjectGroup() now throws CoreException
ProjectUtils: added getAllReferencingProjects() needed for DependencyCache

X-Committed-on-Branch: master