IDE: fix in NED editor: a submodule jumped away on an unpin action.
The temporary position used by the layouter to keep the layout stable was cleared
whenever the hand set position (p tag) changed. This hapenned also when the
p tag was removed (unpin). On unpin action, the temporary position must be kept so
that the module will not be part of the auto-layout process on next UI refresh.

X-Committed-on-Branch: master