IDE: makemake: renamed "compileForDLL" to "forceCompileForDLL"
Makemake actually defines FOO_EXPORT automatically if the
makefile target is a DLL (it does not need this option to be set).

Rather, this option is useful for compiling code that will
eventually end up in a DLL (even if the target of this makefile
is not immediately a DLL). The new name and dialog label expresses
this intent better.

X-Committed-on-Branch: topic/userguide