productizing: IDE source feature fixed.
Each plugin has its own source bundle generated during the IDE build.
The content of the "src" folders are automatically included by the
build process so they must not be marked in the plugin's build.properties
file. The src.includes should be empty (or missing) in the build.properties
files.

These files allow debugging on PDE, but they are not enough to rebuild the
IDE (makefiles, .project etc. are missing). Zipping up ui/ would be abut 10M,
so we do not include it in the distro, instead leave a note that anyone may
ask for it.

X-Committed-on-Branch: (no branch)