gdb setting: set unwindonsignal on
Set unwinding of the stack if a signal is received while in a function
that gdb called in the program being debugged. If set to on,
gdb unwinds the stack it created for the call and restores the context
to what it was before the call. If set to off (the default), gdb stops
in the frame where the signal was received.