opp_run: bugfix; improved error message
On Windows, loadLibErrorMessage printed as random memory garbage

loadLibErrorMessage was a const char*, and it was set to point into an
std::string that was a local variable in oppLoadLibrary()...

    loadLibErrorMessage = opp_getWindowsError(GetLastError()).c_str();

CRASH! when it is printed out later.

Also: the warning message only said "Module cannot be found" but no context
(like, while trying to do what?) or no lib file name!

X-Committed-on-Branch: master