build: fix for #487.
Added $AS_NEEDED_OFF and also fixed a bug where the WHOLE_ARCHIVE_ON was missing
when building a shared lib with gcc. i.e. shared lib models would not be able to
link with an other model that was build statically. (the linker would optimize
out the whole model in the static library)

X-Committed-on-Branch: master