NED: channel name syntax in connections; @defaultname(Foo) attribute for channel types
Syntax:
   r1.pppg++ <--> eth1: EthernetChannel <--> r2.pppg++;
   a.out --> foo: {delay=1ms;} --> b.in;
   a.out --> bar: --> b.in;

Default name may also come from @defaultname of the channel type:

   channel Eth_100M like IEth {
       @defaultname(eth100);
   }

There's a catch with @defaultname though: if the channel type is parametric
in the connection and to be specified with **.liketype, @defaultname
obviously cannot be taken into account, so "channel" is assumed as
channel name. Example:

Assume the above Eth_100M channel, and the connection

   r1.pppg++ <--> <> like IEth <--> r2.pppg++;

Then inifiles:

   **.eth100.typename = "Eth_100M"   # WON'T WORK, BECAUSE "eth100" comes from Eth_100M (Catch-22)
   **.channel.typename = "Eth_100M"  # OK
   **.eth100.datarate = 100.001Mbps  # here "eth100" is OK

Anomaly can be solved by using an explicit channel name in the connection,
or by not using typename lookups (e.g. writing "<someparam> like" instead
of "<> like" is already OK!), or by removing the @defaultname property.

X-Committed-on-Branch: devel