Simulation: initial implementation of ModuleOutputView
Module log messages etc are retrieved by the front-end in the
status HTTP request from Cmdenv. Log is also JSON.

Status is now non-idemportent, because the request also
clears the log in Cmdenv.

In Cmdenv we override several EnvirBase methods to be able to
collect the log.

X-Committed-on-Branch: (no branch)