Simulation: inspector move/resize refactored
Any part of any figure (or SWT widget) can now be used as move/resize
handle; figures/widgets need to be registered in the inspector
container with

    getContainer().addMoveResizeSupport(figure);

and the move/resize operation should be returned by the inspector's

    getDragOperation(figure/canvas, x, y)

method. Rest is done by the inspector container.

X-Committed-on-Branch: anim_tmp_livesim

X-Committed-on-Branch: topic/simfrontend

X-Committed-on-Branch: (no branch)