Simulaton: figures: fixed message layer in CompoundModuleFigure
Message layer was bogus: figures added to it didn't actually appear
on the screen. Turned out this layer was drawn BELOW all
other layers, due to a peculiarity in the addLayerAfter() function.
(It adds layers in order, except the first one which somehow
becomes the last layer!!!)

Fix: replaced bloody addLayerAfter() with simple add() (as we don't
use the useless layer key stuff in LayeredPane anyways). But this
means we need to add layers in reverse order (so, bottom-most first,
topmost last).

X-Committed-on-Branch: anim_experiment

X-Committed-on-Branch: (no branch)