sim: adapted stlwatch classes to cClassDescriptor changes
Since object pointer is no longer arrives in each call, we need
to create a different cClassDescriptor instance for each specialization
of STL containers. (We can no longer figure out the typename inside
the getFieldTypeName() call, so we need to store it in the classdesc
object after we got it in the constructor).

X-Committed-on-Branch: tmp-classdesc