IDE: check in a copy of Eclipse's StyledText widget -- we'll need a modified version for StyledTextHoverControl
The Browser widget is not always available on all platforms (constant
xulrunner problems on Linux, etc), so we want to use StyledText for
presenting HTML. StyledText has all features we need for simple HTML,
only it is not possible to turn on/off line wrapping on line (paragraph)
basis. This missing feature is easy to implement, but we need to
copy StyledText because the affected parts are private or not accessible.

NOTE: THIS VERSION IS ALREADY A LITTLE TWEAKED -- see getWrapWidth()
method (added 'int line' arg & tmp code)

X-Committed-on-Branch: (no branch)