Welcome: revert 87bde6 "Changed default appearance of the main welcome page to be more user friendly."
With OMNEST, Welcome screen is ugly: huge OMNEST logo is placed
in top-left corner with no spacing, and it overlaps with title
of the page.

Must find a way to properly place the logo (at ~2/3rd of page), and then
the new theme can be used again.

X-Committed-on-Branch: omnetpp-4.3.x