logging: simplifications, documentation
- the "_S", static versions of logging macros are no longer needed (explanation below)

- removed the "_P", printf-style logging macros too. They were virtually
  unused, e.g. there was no ev.printf() in the INET Framework, and we
  don't want to encourage them. We can provide an opp_stringf(fmt,...) for
  those who miss it.

- doxygen comments: created a "Logging" group and put everything under that;
  removed doxy comments from classes and macros that we don't want to
  to become public APIs.

About "_S": I implemented something based on the idea from Andy Prowl,
http://stackoverflow.com/questions/15923577/macro-to-return-the-this-pointer-or-null-when-its-not-available,
and added getThisPtr() to cObject and cComponent. The only problem is
that it doesn't work inside a static member of a cObject-derived class;
as a workaround, the user has to add EV_STATICCONTEXT manually.

Using __if_exists could solve it:

   __if_exists(this) {ptr = this;}
   __if_not_exists(this) {ptr = 0;}

but sadly it's Microsoft-only, and there's no equivalent in C++11 or gcc.

X-Committed-on-Branch: (no branch)