ned parser: use separate lex mode (BEGIN(property)) for parsing properties
Assumption: property spans from @ to closing ";", so we can scan that
part of the file differently from the rest.

Ignores special chars (parens, comma, semicolon, equal sign) in nested
context, so grammar doesn't need to deal with it.

This also solves the nested paren problem in property values (any value that
contained parens had to be enclosed in quotes before).

X-Committed-on-Branch: (no branch)