tkenv: fix "Filter window contents" dialog
List of filtered modules could be changed by simply opening and accepting
the dialog without making any change. This occurs to subtrees that
are never expanded while the dialog is open. (i.e. reproducing the
bug needed compound modules to be present in the simulation, e.g.
the "routing" example)

X-Committed-on-Branch: omnetpp-4.3.x