sim: added emit(simsignal_t, const cObject*) to cComponent
This was necessary because otherwise the compiler casts the const pointer
to long (!!!) and calls that emit().

The new method delegates to the const-less version, which is not very
clean. It would be better to do the other way round (let non-const delegate
to the const version), however, for that the listeners would have to
accept const pointer which would be a painful change.

X-Committed-on-Branch: zoli