Fixed bug 716.
Action contributors are used by both the eclipse framework to create buttons on toolbars and
by our widgets to create the popup menus. We need to make sure that the stateful actions of
the popup menus and the toolbars are in sync. We achieve this goal by using the very same
instance of the contributors and the actions they contribute. This has to work with both
editors and views.