tkenv: info bar removed from toplevel inspectors, and added to embedded GenericObjectInspector
For toplevel inspectors, the window title already contains the same info
(less the icon, bt it's not that important).

ALSO, removed the "number of items" status label from inspector listbox, it will
be displayed separately (e.g. in the label of the "Contents" tab)

X-Committed-on-Branch: topic/tkenv

X-Committed-on-Branch: (no branch, rebasing topic/tkenv)