nedxml: fix #773: ned2.lex/msg2.lex: newline inside properties confused the parser
To reproduce, process the following NED file:

@f(a=
foo);

with: nedtool -x a.ned

In the resulting XML, the property value is garbage instead of "foo".

It was a bug in the lexer, ned2.lex (and also msg2.lex):
in Flex, "." doesn't match the newline, one must explicitly handle
that case (or use {S} that includes \n too).

X-Committed-on-Branch: topic/ide-graphics-5.0