ide: NED editor refactoring: be prepared of Tkenv-like zoom
Instead of zooming the whole editor contents via ScalableGraphics,
we'll want zoom to change the pixel/unit ratio within the selected
compound module. For this, "scale" (== the new zoom level) has been
added to CompoundModuleEditPart.

This also made it necessary to store positions and sizes in floats. GEF's
Point and Rectangle are not suitable, PrecisionPoint/Rect are convoluted,
so we created PointF/DimensionF/RectangleF. This is also useful because
now model and pixel coordinates are set apart also by the Java type system.