test: Fixed a number test case output checks.
When the simulation is run in express mode, the first/last printed event
number must be less by 1. When the simulation is run in non-express mode,
the 'simulation ended at event #x' must be less by 1. These are changed due
to the changes where the event number gets incremented in cSimulation.