qtenv: Fixed bug related to showing array fields.
When the number of elements change in an array field, it crashed the user
interface in an ASSERT during collecting the expanded tree nodes. The reason
is that the underlying data structure changed while the tree nodes represent
the old state. This wouldn't be a problem, since the expanded tree nodes is
just a set of strings, if hasChildren and getChildCount would be consistent.
This patch fixes that for the case when the tree node is filled.