osg: fix: earth scenes did not display in the osg-intro sample
Root cause: zNear/zFar was not set explicitly in these demos, so it stuck
with the default values in cOsgCanvas: 1 and 1000.  This caused basically
all of the scene to be clipped away (out of z range).

Solution: cOsgCanvas does NOT contain any zNear/zFar by default, only when
explicitly set. In the viewer, activate near/far autocompute mode when
cOsgCanvas has no zNear/zFar, otherwise set DO_NOT_COMPUTE_NEAR_FAR.

Adjusting the aspect ratio in OsgViewer::resizeEvent() is not needed when
setProjectionResizePolicy(HORIZONTAL) is set.