sim: fix crash when inspecting the owner of HistogramRecorder's inner cStatitistic
Reason for the crash: when mapping the value returned by getOwner()
to HistogramRecorder, the pointer value will be off (will point to
the cObject part instead of the start of the object).

This cannot be easily fixed, as multiple inheritance is not mapped well
into cClassDescriptors.

Workaround: StatisticsRecorder no longer owns the cStatitistic object
(but calls removeFromOwnershipTree() on it instead) so it no longer
needs to be a cObject.