common,layout: Disallow invalid LCGRandom seeds, layout scaling fix.
In rare cases, during the prelayouting phase of the advanced layouter,
when scaling the layout to make it fill the given bounding box, if a node
had a size equivalent to the bounding box in any dimension, a division
by zero caused unwanted NaN values, which later triggered the failure of
an assertion which required the distance between two nodes to be greater
than zero (but it would be NaN in this case). Discovered by the EnvirApi test.