doc,ide: removing "bgl" display string tag, step 1
The "bgl" could be used to specify the layout seed for the compound module.
It has not been supported by Qtenv, as it was unclear how it was
supposed to coexist with Qtenv's "Relayout" button. We are now removing
it altogether.

This commit removes mentions from the IDE and the documentation.