qtenv: Perform holding animations at the end too when animating until the next event.
This should fix the ASSERT failure in chopHoldingTail (from > currentPart).
Also do not move the next event marker until they are done.
For example in the routing example, the final part of the animation
between to nodes (inside the receiving node) is a holding one, but
follows a nonholding anim. To truly arrive right before delivery,
we have to do that too, right after the first two parts.