qtenv: Don't strip namespaces for internal uses.
Like storing inspector windows and preferences, and appending to the message filter.
This should fix the bug when the "exclude messages like this from animation"
appended to the filter with the class name stripped according to the user setting
in the preferences dialog, but the filter engine did not accept it that way.
Also a small, now redundant idMatches function was removed.