qtenv: OSG Viewer is in a separater library, loaded at runtime. Comments below.
Motivation:

OSG and especially osgEarth added a huge number of dependencies to Qtenv,
and to all simulation executables independent of whether or not 3D was actually
used by the simulation. This increased load time and made debugging harder.

Change:

Qtenv and the sim library does not depend at all on OSG or osgEarth.
OSG support is in a separate library loaded at runtime.
If the model uses OSG, then it needs to link with the necessary libraries.

Sim API change:

One must use cOsgCanvas::EarthViewpoint instead of osgEarth::Viewpoint in models.

Details:

The library is called oppqtenv-osg. It is looked for in the same places as the
other libraries specified by the -l command line argument of opp_run.
It is loaded with dlopen/LoadLibrary, so the rules of their operation apply.

OSG frame rate was decreased from 60 to 30 fps.

If the library could not be loaded, dummy viewers will be created in place
of the real ones. The dummy viewers just print a short message about the
failed loading of the library.
The class descriptor for cOsgCanvas, its inner classes, and for osg::Node
are included in/loaded with this support library.

No WITH_OSG or WITH_OSGEARTH conditionals anywhere in omnetpp or qtenv.
The only difference with WITH_OSG is whether or not the support library
gets built. Inside that, there are WITH_OSGEARTH conditionals still.