launch: fixes: typo "exisits" (STILL!); fix misleading error message on build error
Nothing indicates that there was a build error, so user was left confused.
Ideally, we'd cancel the launch of the build failed, but it's apparently
not possible with Eclipse (or not obvious how to do it). At least now
the message suggests that maybe the failed build is the cause...