cmdenv: Prevent a division by zero causing a buffer overflow.
If the simtimeLimit was set to zero (which happened in some INET
fingerprint tests, where the "--sim-time-limit=0s" command line
argument was supplied to opp_run), 'simtimeRatio' became NaN
as a result of the division by zero. Then 'ratio' and 'totalRatio'
became NaN as well. When casted to int, the result was some large
negative number (-2147483648). The final effect is that we printed
47 bytes into a 32 byte long buffer.

The size of 'buf' was just large enough to fit the 100% twice.