tests: updated fingerprint test script
Use full command line in the CSV file. The second column of the CSV now
specifies the full command line to be executed as a test. If the cmdLine
starts with an option (i.e. no program name specifies), then the global
executable name is used (use the -e option to specify the command globally)

X-Committed-on-Branch: topic/fingerprinttest