sim: cDefault/AutoHistogramStrategy: fix std::bad_alloc on collecting an extremely large value
To reproduce, collect e.g. 1e20 when existing hist range is "normal", e.g. 0..10.

extendBinsTo() first wanted to add a (nearly) infinite number of blank
bins to include "value" in the histogram range before reducing the
number of bins to targetNumBins.

Solution: detect the case when the new bin size would be greater
than the whole existing histogram, and handle that specially.