layout: Fix "Advanced" layouter not doing anything with only anchored nodes.
In one example network, the samples/routing/RandomMesh, there are only
anchored "nodes" (submodules) - no fixed and no independently movable
nodes - since the whole network is one matrix with undefined position.
In this network, when the "Advanced" layouter way selected, none of
the submodules got a position after layouting. This was most likely
never tested because the auto selection chooses the simple layouter
in this simulation.