sim: Add infinity counting to cAbstractHistogram and subclasses.
Collecting positive/negative infinity no longer breaks the histogram.
Instead, infinities are counted as overflows/underflows, and the
bins are set up as if the infinities were not there.

This also changes the ad-hoc text file format used by saveToFile()/
loadFromFile(), because the classes need to keep the number of
finite and infinite overflows, finite extrema, etc. in addition
to data kept so far.