qtenv: Fix a crash when closing inspector windows with Qt >=5.12.
Before, the inspector was deleted by us manually, right inside the
closeEvent() callback. With older Qt versions, this caused no problems,
because no widget member was accessed after the event.
Not so long ago this changed, so we now have to rely on Qt for
deleting the Inspector when it is closed, but it works well.

See: https://bugreports.qt.io/browse/QTBUG-78314