test/core: adjusted expected test output
In ini files lines that set module parameters, the value can be an expression,
so string literals must be quoted in order to be able to distinguish between
things like 1+1 and "1+1", true and "true", or ""+"" or "\"\"+\"\"".

Values must be saved in the result files in the same way, that is,
strings must be quoted.

Here, "typename" counts as parameter too, as its value can be an
expression.

Since result file readers (e.g. the IDE Analysis Tool) remove and discard
one layer of quotes if present, values need to appear double-quoted in
the result file so that the inner quotes appear as part of the value.