sim,envir: let xmldoc() interpret the file name as relative to the dir the expression comes from
That is, when xmldoc() occurs in an included ini file, the file name
is now interpreted as relative to the directory containing the included
ini file (as opposed to being relative to the main ini file or the
working directory.)

If xmldoc() occurs in a NED file, thr file name is relative to the directory
where that NED file was loaded from.

In addition to updating the implementationi of xmldoc(), it had to be
ensured that the function gets access to the directory where the ini
entry or NED expression occurred (baseDirectory). This directory was
added to the evalation context (cExpression::Context).

cNedFunction (and thus, Define_NED_Function()) was updated to accept
an alternative function signature which accepts cExpression::Context
instead of cComponent* as context argument.

getSourceFileDirectory() method was added to cComponentType.