utils: opp_featuretool: make it complain less about fixable problems
(Auto repair made the default mode of operation.)

In general, the tool has become more forgiving: missing .featurestates
file and extra/missing entries in it don't cause an error or even
a message. Missing file and file entries are initialized with the default
enablement state; extra entries are preserved (in the hope they'll be
useful, e.g. after switching to a different topic branch.

The content of .nedexclusions file is also adjusted (without stopping
with an error).