envir: added support for result directory subdivision: $iterationvarsd, resultdir-subdivision config option
Many file managers and other tools (inc. Eclipse) break down if there
are too many files in single directory (10,000 or more). The common
workaround is to group the files into a subdirectory tree. (For example,
git also uses this technique to store the contents of .git/objects dir).

In OMNeT++, this problem occurs when a parameter study produces too many
result files. To solve it, it is natural to use the iteration variables to define
the directory hierarchy. This commit does it in the following way:

1. Defines a "$iterationvarsd" variable (run attribute), which is similar
   to  "$iterationvarsf", only it contains '/' instead of ','.

2. Turning on the new "resultdir-subdivision" config option will cause
   "$iterationvarsd" to be appended to the name of the results directory.
   This way, files will be created inside a directory tree defined by
   the iteration variables.

When the user wants set up the directory hierarchy in a different way,
thay can do so by setting the "results-dir" config option and appending
stuff to th value. E.g.: results-dir = "results/${repetition}"