ide/scave: remove inputName from loadFile() arg list
Issue: when loadFile() detects that a file is already loaded and up to
date, but its inputName field is different, should it update?
In the IDE, it should, but this decision should not be built into
the API.

So in the new code, we call setFileInput(file,inputName) separately
after the loadFile() call.