scave: IDList: make IDList immutable (remove mutator member functions)
Motivation: it makes it easier to write correct code in Java.

IDLists are often cached at various places in Java, in the assumption
that its contents never changes. Now it is guaranteed.

An op=() still exists in C++ for technical reasons, but it's not
accessible from Java.

For building IDLists one should use std::vector<ID>, which goes by the
name IDListBuffer in Java.