scave: proper "field scalars"
Before, we stored the fields of statistics/histograms/vector as proper
scalars as well, with an "isField" bit set, in addition to inside
StatisticsResult/VectorResult items.

This patch removes this reduncancy of storage. Instead, we extend IDs
so that they represent a scalar as "field x or statistics or vector y".
In getItem() and getScalar() of such an ID, we fabricate a temporary
ScalarResult object, and return that by value.

One trick to simplify usage is to pass a "buffer" ScalarResult into
getItem()/getScalar(), so it can return a simple pointer no matter
whether the ID identifies a field scalar or a normal result item.