ide/scave: Content assist on the Advanced tabs of chart dialogs
For rcParams, the input comes from Python. First time we execute
a chart script, we do an extra call to return the result of a
"mpl.rcParams()" call to the IDE. This is remembered, and used
in content assist.

NOTE: using the Enter key to accept proposals currently doesn't work,
only the mouse. This is an error in the adapter.