ide/scave: remove redundant classes, #1
Sorter: originally used (probably) for the Inputs page. Currently it was
only used for sorting the list of runs within an result file (Inputs page).
This use was removed, as runs in a result file are best shown in their
natural order.

ResultItemFormatter: this kind of stuff is nowadays done in Python.

ScaveModelLabelDecorator: originally used for displaying error icons
on bogus files in the Inputs view. The Inputs view no longer shows
files we couldn't load, so this class has no use.