ide/scave: Speed up temp chart opening by bypassing pass filter evaluation
Instead, we pass the selection IDList by means of a filterExpr->IDList cache.

This provides tremendous performance improvement for temp charts,
as the filter doesn't need to be run on the full set of loaded results.

The IDList selected on "Browse Data" page is just "tunnelled through" to
ResultProvider, via a secondary map in MemoizationCache.

Also: ResultSelectionFilterGenerator: rename DumbFilter to QuickFilter,
and make it less dumb (factor out those field that are the same among
all results).