ide/common: ContentAssistUtil: fix content assist popup position for multi-line Text widgets
The proposal popup came up alongside the bottom edge of the text control,
not below the edited line.

The solution is to add code configureText() that adjusts the position
from a listener. This is the copy of the same code in configureStyledText().

It is possible that this code should really be part of ContentProposalAdapter
in some form.

THE ONLY PROBLEM with Text widgets is that we don't have Undo/Redo
support for them. (For StyledText we do.)