ide/common: MatchExpressionContentProposalProvider rewritten, for scalability, bugfixing and more practical API
The old code failed with StackOverflowException on long inputs
(10,000+ clauses) that we have in Scave, due to the use of recursive
descent parsing.

For content assist with a grammar as simple as as this one
we don't need parsing, knowing the last few tokens in enough.

This is exactly what this class does. We don't even tokenize the whole
string because even that can be expensive (a few hundred millisecods and up),
only the last few hundred chars. (We scan thr whole string to find a good
starting point for the tokenizing though, like we can't start within a
string constant because the whole result will be off.)

The API is now getFieldNameProposals(), getDefaultFieldValueProposals()
and getFieldValueProposals() (user needs to provide these methods),
which is MILES better than original "here's the parse tree, do whatever
you want with it".