ide/scave: DataTable/LargeTable: selection-related fixes
DataTable.setIDList(): Clear selection. It is invalid with the new IDList so it
cannot be left as is, and preserving it is too complicated because of
change listener hell. Plus no user cares if selection is lost after
changing the filters, really.

setItemCount(): also clear the selection, for the same reason.

Also: removed getSelectedItem() as it is broken concept, because there is multiple selection.
Use focus item instead, at that 1 place where it is needed.

SetFilterBySelectedCellAction: make it slightly less broken.