sim: send()/sendDirect: introduced SendOptions and tx updates; cChannel API and cDatarateChannel changes
Note: substantial changes in this patch are concentrated in
cSimpleModule and cDatarateChannel, with necessary extra fields
added to cPacket. The rest are to follow signature changes in
cEnvir and cChannel API methods, and to record the extra fields
in eventlog files.

Updating a transmission (aka tx updates):

Initial send() is interpreted as: "packet transmission begins now,
packet content and duration are, as things are now, going to be this".

Following that, an "update" (or any number of updates) can be sent.
It is a packet with the updated ("actual") content, and with a
"remaining transmission duration" attached. Updates may only
be sent while transmission is still ongoing (t < txEndTime; note
it is strictly less, not less-or-equal!)

As an example, aborting a transmission is done by sending a
packet with a truncated content and a remaining duration of zero.

The transmission it modifies (in case there are parallel transmissions
on the channel) is identified by the packet ID of the original
transmission. Channels should understand updates and handle them
accordingly. (Change to cDatarateChannel is part of this patch.)

Receivers that receive the packet at the end of the reception (default)
will only receive the final update, the original packet and
intermediate updates are absorbed by the simulation kernel.

Receivers that receive the packet at the start of the reception (see
cGate::setDeliverImmediately(), ex setDeliverOnReceptionStart())
should be prepared to receive all of the original packet and the
updates, and handle them appropriately. Tx updates can be recognized
from cPacket::isUpdate() returning true. cPacket::getRemainingDuration()
returns the remaining transmission duration, and cPacket::getDuration()
the total transmission duration.

As a safeguard against unprepared modules accidentally processing
tx updates as normal full-blown packets, the module is only given
tx updates if it explicitly declares that it is able to handle them
properly. The latter is done by the module calling
setTxUpdateSupport(true) e.g. in initialize().

Non-transmission channels don't handle updates any different from
other messages and packets.

SendOptions was introduced as a means to handle combinatorial explosion
of send() variants.

More changes:
- send() and sendDirect() now accept a SendOptions; sendDelayed() and
  the sendDirect() variant with delay and duration now convert their
  extra args to a SendOptions, and delegate to the "standard"
  send()/sendDirect() versions.

- cDatarateChannel: allow overriding the duration computed from the
  channel datarate from the send() call.

- cDatarateChannel: non-packet messages pass through without
  interfering with packets.

- cDatarateChannel: disabled channels to let transmission updates through,
  so that it is possible for the transmitter module to abort the
  ongoing packet transmission.

- Note: tx updates (without duration/remainingDuration) are allowed
  on paths without transmission channels.

- signature change in methods that participate in the send()/sendDirect()
  protocol: SendOptions was added to the arg list. This affects:
  cGate::deliver(), cModule::arrived(), cChannel::processMessage()

- in cChannel::processMessage(), result_t was renamed cChannel::Result,
  and it is now a proper return value (not an output parameter);

- remainingDuration was added to cChannel::Result.

- cDatarateChannel's datarate is now optional: set it to 0 or nan
  to leave it unspecified.  This change was necessary to support
  transmitting frames with per-frame data rate selection.

- Removed forceTransmissionFinishTime() from cChannel (and esp. cDatarateChannel)
  It was a hack. Now you can use SendOptions::updateTx() instead.

- Rename: setDeliverOnReceptionStart() -> setDeliverImmediately()

- Added cSimpleModule::supportsTxUpdates() flag

- cPacket now carries a remainingDuration field

- cPacket: eliminated FL_ISRECEPTIONSTART; isReceptionStart() now uses
  remainingDuration as input; added a similar isReceptionEnd() method.

- cPacket::str() overhaul to reflect new fields and uses

- in the APIs, sendDelay and propagationDelay (which were sort of
  combined into a single value) are now distinct, and handled
  separately.

- eventlog format changes: BeginSendEntry: added sendDelay and
  origPacketId; SendHopEntry, SendDirectEntry: added remainingDuration

- cSimpleModule: factored out a throwNotOwnerOfMessage() method to
  eliminate a large chunk of nearly identical code in send() and
  sendDirect().

- cEnvir callbacks also changed: beginSend() takes SendOptions;
  messageSendHop() and messageSendDirect() take cChannel::Result
  instead of separate args.

Note: cChannel::Result became a toplevel class (ChannelResult) due
to the need to break circular dependency: "cenvir.h" cannot include
"cchannel.h" (circularity) and C++ does not allow inner classes
to be forward-declared (why??).